+++ lib/unisec.rb
@@ -4,6 +4,7 @@ require 'unisec/version'
 
 require 'unisec/bidi'
 require 'unisec/confusables'
+require 'unisec/decdump'
 require 'unisec/hexdump'
 require 'unisec/normalization'
 require 'unisec/properties'